feat: move appearance + sharing mode to settings sub-pages
