+++ lib/screens/settings/settings_page.dart
@@ -46,6 +46,8 @@ import 'package:grid_frontend/screens/settings/developer_tools_screen.dart';
 import 'package:grid_frontend/screens/settings/encryption_keys_screen.dart';
 import 'package:grid_frontend/screens/settings/home_location_picker_screen.dart';
 import 'package:grid_frontend/screens/settings/profile_photo_screen.dart';
+import 'package:grid_frontend/screens/settings/appearance_settings_screen.dart';
+import 'package:grid_frontend/screens/settings/sharing_mode_screen.dart';
 import 'dart:io' show Platform;
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -2400,40 +2402,21 @@ class _SettingsPageState extends State<SettingsPage> {
 
             // ── Appearance ──────────────────────────────────
             const GridSectionHeader(text: 'Appearance'),
-            AnimatedBuilder(
-              animation: ThemeController.instance,
-              builder: (context, _) => _buildSectionCard(
-                theme: theme,
-                colorScheme: colorScheme,
-                children: [
-                  _ThemeModeRow(
-                    icon: Icons.brightness_auto_outlined,
-                    title: 'System',
-                    selected:
-                        ThemeController.instance.mode == ThemeMode.system,
-                    onTap: () => ThemeController.instance
-                        .setMode(ThemeMode.system),
+            _buildSectionCard(
+              theme: theme,
+              colorScheme: colorScheme,
+              children: [
+                AnimatedBuilder(
+                  animation: ThemeController.instance,
+                  builder: (context, _) => _buildInfoRow(
+                    icon: Icons.brightness_6_outlined,
+                    title: 'Appearance',
+                    value: _themeModeLabel(ThemeController.instance.mode),
+                    onTap: _openAppearance,
+                    colorScheme: colorScheme,
                   ),
-                  _buildSettingsDivider(),
-                  _ThemeModeRow(
-                    icon: Icons.light_mode_outlined,
-                    title: 'Light',
-                    selected:
-                        ThemeController.instance.mode == ThemeMode.light,
-                    onTap: () => ThemeController.instance
-                        .setMode(ThemeMode.light),
-                  ),
-                  _buildSettingsDivider(),
-                  _ThemeModeRow(
-                    icon: Icons.dark_mode_outlined,
-                    title: 'Dark',
-                    selected:
-                        ThemeController.instance.mode == ThemeMode.dark,
-                    onTap: () => ThemeController.instance
-                        .setMode(ThemeMode.dark),
-                  ),
-                ],
-              ),
+                ),
+              ],
             ),
 
             // ── Sharing ─────────────────────────────────────
@@ -2452,9 +2435,12 @@ class _SettingsPageState extends State<SettingsPage> {
                   colorScheme: colorScheme,
                 ),
                 _buildSettingsDivider(),
-                _SharingModeRow(
-                  selected: _sharingMode,
-                  onChanged: _setSharingMode,
+                _buildInfoRow(
+                  icon: Icons.tune_outlined,
+                  title: 'Sharing mode',
+                  value: _sharingModeLabel(_sharingMode),
+                  onTap: _openSharingMode,
+                  colorScheme: colorScheme,
                 ),
                 _buildSettingsDivider(),
                 _buildToggleOption(
@@ -2657,6 +2643,37 @@ class _SettingsPageState extends State<SettingsPage> {
       ),
     );
   }
+
+  void _openAppearance() {
+    Navigator.push(
+      context,
+      MaterialPageRoute(builder: (_) => const AppearanceSettingsScreen()),
+    );
+  }
+
+  void _openSharingMode() {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (_) => SharingModeScreen(
+          initial: _sharingMode,
+          onChanged: _setSharingMode,
+        ),
+      ),
+    );
+  }
+
+  String _themeModeLabel(ThemeMode m) => switch (m) {
+        ThemeMode.system => 'System',
+        ThemeMode.light => 'Light',
+        ThemeMode.dark => 'Dark',
+      };
+
+  String _sharingModeLabel(SharingMode m) => switch (m) {
+        SharingMode.light => 'Light',
+        SharingMode.balanced => 'Balanced',
+        SharingMode.live => 'Live',
+      };
 
   void _openProfilePhoto() {
     final client = Provider.of<Client>(context, listen: false);
@@ -3900,264 +3917,3 @@ class _DangerBullet {
   final String text;
 }
 
-/// The "Sharing mode" row inside the Sharing card. Renders a 3-position
-/// segmented control plus a description of the selected mode (what
-/// friends see + the battery estimate). Designed so the selected mode
-/// owns the screen without forcing the user to read all three.
-class _SharingModeRow extends StatelessWidget {
-  const _SharingModeRow({required this.selected, required this.onChanged});
-
-  final SharingMode selected;
-  final ValueChanged<SharingMode> onChanged;
-
-  static const _modes = <_ModeInfo>[
-    _ModeInfo(
-      mode: SharingMode.light,
-      label: 'Light',
-      icon: Icons.bedtime_outlined,
-      tagline: 'Updates only when you change places.',
-      friendsSee: 'Friends see when you arrive somewhere new.',
-      battery: '~1% per day',
-    ),
-    _ModeInfo(
-      mode: SharingMode.balanced,
-      label: 'Balanced',
-      icon: Icons.balance_outlined,
-      tagline: 'Updates every ~60s while moving.',
-      friendsSee: 'Your dot lags real life by 1–2 minutes.',
-      battery: '~2–4% per day',
-    ),
-    _ModeInfo(
-      mode: SharingMode.live,
-      label: 'Live',
-      icon: Icons.gps_fixed_rounded,
-      tagline: 'Updates every ~30s while driving.',
-      friendsSee: 'Navigation-grade. For trips.',
-      battery: '~5–8% per day',
-    ),
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    final info = _modes.firstWhere((m) => m.mode == selected);
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16, 14, 16, 14),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Icon(
-                Icons.tune_rounded,
-                size: 18,
-                color: context.gridColors.text2,
-              ),
-              const SizedBox(width: 10),
-              Text(
-                'Sharing mode',
-                style: GoogleFonts.getFont(
-                  'Geist',
-                  fontSize: 15,
-                  fontWeight: FontWeight.w600,
-                  letterSpacing: -0.01,
-                  color: context.gridColors.text,
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 12),
-          Container(
-            padding: const EdgeInsets.all(4),
-            decoration: BoxDecoration(
-              color: context.gridColors.surface2,
-              borderRadius: BorderRadius.circular(999),
-              border: Border.all(color: context.gridColors.hairline),
-            ),
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                for (final m in _modes)
-                  Expanded(
-                    child: _SharingModeChip(
-                      info: m,
-                      active: m.mode == selected,
-                      onTap: () => onChanged(m.mode),
-                    ),
-                  ),
-              ],
-            ),
-          ),
-          const SizedBox(height: 12),
-          Text(
-            info.tagline,
-            style: GoogleFonts.getFont(
-              'Geist',
-              fontSize: 13.5,
-              fontWeight: FontWeight.w500,
-              color: context.gridColors.text,
-              letterSpacing: -0.005,
-            ),
-          ),
-          const SizedBox(height: 2),
-          Text(
-            info.friendsSee,
-            style: GoogleFonts.getFont(
-              'Geist',
-              fontSize: 12.5,
-              color: context.gridColors.text2,
-              height: 1.4,
-            ),
-          ),
-          const SizedBox(height: 6),
-          Row(
-            children: [
-              Icon(
-                Icons.battery_5_bar_rounded,
-                size: 13,
-                color: context.gridColors.mint,
-              ),
-              const SizedBox(width: 6),
-              GridMono(
-                info.battery,
-                size: 10.5,
-                letterSpacing: 0.08,
-                color: context.gridColors.mint,
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _SharingModeChip extends StatelessWidget {
-  const _SharingModeChip({
-    required this.info,
-    required this.active,
-    required this.onTap,
-  });
-
-  final _ModeInfo info;
-  final bool active;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return Material(
-      color: Colors.transparent,
-      child: InkWell(
-        onTap: onTap,
-        borderRadius: BorderRadius.circular(999),
-        child: Container(
-          height: 36,
-          alignment: Alignment.center,
-          decoration: BoxDecoration(
-            color: active ? context.gridColors.mintFaint : Colors.transparent,
-            borderRadius: BorderRadius.circular(999),
-            border: active
-                ? Border.all(color: context.gridColors.mintSoft, width: 1)
-                : null,
-          ),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Icon(
-                info.icon,
-                size: 14,
-                color: active ? context.gridColors.mint : context.gridColors.text3,
-              ),
-              const SizedBox(width: 6),
-              Text(
-                info.label,
-                style: GoogleFonts.getFont(
-                  'Geist',
-                  fontSize: 13,
-                  fontWeight: FontWeight.w600,
-                  letterSpacing: -0.005,
-                  color: active ? context.gridColors.mint : context.gridColors.text2,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ModeInfo {
-  const _ModeInfo({
-    required this.mode,
-    required this.label,
-    required this.icon,
-    required this.tagline,
-    required this.friendsSee,
-    required this.battery,
-  });
-
-  final SharingMode mode;
-  final String label;
-  final IconData icon;
-  final String tagline;
-  final String friendsSee;
-  final String battery;
-}
-
-/// Row used by the Appearance section: shows an icon + label and a mint check
-/// when selected, with a mint-faint background tint highlighting the active
-/// option. Mirrors the menu/info row paddings used elsewhere in this file.
-class _ThemeModeRow extends StatelessWidget {
-  const _ThemeModeRow({
-    required this.icon,
-    required this.title,
-    required this.selected,
-    required this.onTap,
-  });
-
-  final IconData icon;
-  final String title;
-  final bool selected;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    final Color iconColor =
-        selected ? context.gridColors.mint : context.gridColors.text2;
-    final Color titleColor = context.gridColors.text;
-
-    return Material(
-      color: selected ? context.gridColors.mintFaint : Colors.transparent,
-      child: InkWell(
-        onTap: onTap,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 14),
-          child: Row(
-            children: [
-              Icon(icon, size: 20, color: iconColor),
-              const SizedBox(width: 16),
-              Expanded(
-                child: Text(
-                  title,
-                  style: GoogleFonts.getFont(
-                    'Geist',
-                    color: titleColor,
-                    fontSize: 15,
-                    fontWeight: FontWeight.w500,
-                    letterSpacing: -0.01,
-                  ),
-                ),
-              ),
-              if (selected)
-                Icon(
-                  Icons.check_rounded,
-                  size: 18,
-                  color: context.gridColors.mint,
-                ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
